+++ model/lib/src/fetch.dart
@@ -75,6 +75,9 @@ final class MetaFetch {
   static MetaFetch forTest() => MetaFetch._(true);
 
   /// Define a value of user agent when making request in [fetchFromHttp].
+  /// 
+  /// If [disguiseUserAgent] enabled, any changes will not be applied
+  /// until it disabled and uses user defined again.
   static void changeUserAgent(
       [String userAgent = OgHrefClient.DEFAULT_USER_AGENT_STRING]) {
     OgHrefClient.userAgent = userAgent;
@@ -94,10 +97,14 @@ final class MetaFetch {
 
   /// Determine using user agent from web browser instead of [userAgentString].
   ///
-  /// This does not applied under native environment.
+  /// This does not applied under native environment and no returned value
+  /// changes of [userAgentString].
   static bool get disguiseUserAgent => OgHrefClient.disguise;
 
   /// Retrive current preference of user agent [String].
+  /// 
+  /// When [disguiseUserAgent] enabled in web platform, it returns web
+  /// browser's user agent instead of user defined value.
   static String get userAgentString => OgHrefClient.userAgent;
 
   /// Retrive current preference of timeout.